move ExpensesInfo model
