+++ src/app/fyle/merge-expense/merge-expense.page.ts
@@ -26,7 +26,7 @@ import { CustomFieldsService } from 'src/app/core/services/custom-fields.service
 import { CustomInputsService } from 'src/app/core/services/custom-inputs.service';
 import { DependentFieldsService } from 'src/app/core/services/dependent-fields.service';
 import { ExpenseFieldsService } from 'src/app/core/services/expense-fields.service';
-import { ExpensesInfo } from 'src/app/core/services/expenses-info.model';
+import { ExpensesInfo } from 'src/app/core/models/expenses-info.model';
 import { MergeExpensesService } from 'src/app/core/services/merge-expenses.service';
 import { SnackbarPropertiesService } from 'src/app/core/services/snackbar-properties.service';
 import { TrackingService } from 'src/app/core/services/tracking.service';
